websocket_client/requirements.txt
pandas==2.2.2
jiwer==3.0.4
websocket-client==1.8.0
httpx==0.27.0
requests==2.32.3
starlette==0.45.2
pandas==2.2.2
